Controller Advice for 404 (#26)
* Controller Advice for 404

* Controller Advice for 404 - Service tests

* Author Update - Controller Advice 404

* Author Delete - Controller Advice 404

* Author Delete - Sonar Issues Fix

* Suppress curly bracket checkstyle

+++ src/main/java/mr/cookie/spring6udemy/services/AuthorServiceImpl.java
@@ -4,8 +4,8 @@ import lombok.RequiredArgsConstructor;
 import mr.cookie.spring6udemy.model.mappers.AuthorMapper;
 import mr.cookie.spring6udemy.model.model.Author;
 import mr.cookie.spring6udemy.repositories.AuthorRepository;
+import mr.cookie.spring6udemy.services.exceptions.NotFoundEntityException;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Service;
 
@@ -32,12 +32,12 @@ public class AuthorServiceImpl implements AuthorService {
                 .orElse(Collections.emptyList());
     }
 
-    @Nullable
+    @NotNull
     @Override
     public Author findById(long id) {
         return this.authorRepository.findById(id)
                 .map(this.authorMapper::map)
-                .orElse(null);
+                .orElseThrow(() -> new NotFoundEntityException(id, Author.class));
     }
 
     @Override
@@ -53,7 +53,7 @@ public class AuthorServiceImpl implements AuthorService {
     @Override
     public @NotNull Author update(long id, @NotNull Author author) {
         var existingDto = this.authorRepository.findById(id)
-                .orElseThrow();
+                .orElseThrow(() -> new NotFoundEntityException(id, Author.class));
 
         existingDto.setFirstName(author.getFirstName());
         existingDto.setLastName(author.getLastName());
@@ -66,7 +66,14 @@ public class AuthorServiceImpl implements AuthorService {
 
     @Override
     public void deleteById(long id) {
-        this.authorRepository.deleteById(id);
+        var doesAuthorExist = this.authorRepository.findById(id)
+                .isPresent();
+
+        if (doesAuthorExist) {
+            this.authorRepository.deleteById(id);
+        } else {
+            throw new NotFoundEntityException(id, Author.class);
+        }
     }
 
 }
